+++ packages/ui/CHANGELOG.md
@@ -7,10 +7,14 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Changed
 
 - `ui/Marquee` to reduce redundant lifecycle work during high-frequency re-renders
+### Added
+
+- `ui/resolution` Added support for linear scaling type
 
 ### Fixed
 
-- `ui/VirtualList.VirtualList`, to adjust itemSize properly when resizing window
+- `ui/VirtualList.VirtualList` to adjust itemSize properly when resizing window
+- `ui/VirtualList.VirtualList` to scroll properly when another VirtualList exist
 
 ## [5.4.2] - 2026-01-28